requirements.txt
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via pip-tools
idna==2.10                # via requests
pip-tools==5.2.1          # via -r requirements.in
python-dateutil==2.8.1    # via -r requirements.in
requests==2.24.0          # via -r requirements.in
six==1.15.0               # via pip-tools, python-dateutil
urllib3==1.25.9           # via requests

# The following packages are considered to be unsafe in a requirements file:
# pip
